refactor(combat): reorganize GameEngine and CombatScreen components for improved structure and clarity
